+++ tests/api/test_product.py
@@ -104,6 +104,69 @@ QUERY_FETCH_ALL_PRODUCTS = """
 """
 
 
+QUERY_PRODUCT = """
+    query ($id: ID, $slug: String){
+        product(
+            id: $id,
+            slug: $slug,
+        ) {
+            id
+            name
+        }
+    }
+    """
+
+
+def test_product_query_by_id(
+    user_api_client, product,
+):
+    variables = {"id": graphene.Node.to_global_id("Product", product.pk)}
+
+    response = user_api_client.post_graphql(QUERY_PRODUCT, variables=variables)
+    content = get_graphql_content(response)
+    collection_data = content["data"]["product"]
+    assert collection_data is not None
+    assert collection_data["name"] == product.name
+
+
+def test_product_query_by_slug(
+    user_api_client, product,
+):
+    variables = {"slug": product.slug}
+    response = user_api_client.post_graphql(QUERY_PRODUCT, variables=variables)
+    content = get_graphql_content(response)
+    collection_data = content["data"]["product"]
+    assert collection_data is not None
+    assert collection_data["name"] == product.name
+
+
+def test_product_query_error_when_id_and_slug_provided(
+    user_api_client, product, graphql_log_handler,
+):
+    variables = {
+        "id": graphene.Node.to_global_id("Product", product.pk),
+        "slug": product.slug,
+    }
+    response = user_api_client.post_graphql(QUERY_PRODUCT, variables=variables)
+    assert graphql_log_handler.messages == [
+        "saleor.graphql.errors.handled[ERROR].GraphQLError"
+    ]
+    content = get_graphql_content(response, ignore_errors=True)
+    assert len(content["errors"]) == 1
+
+
+def test_product_query_error_when_no_param(
+    user_api_client, product, graphql_log_handler,
+):
+    variables = {}
+    response = user_api_client.post_graphql(QUERY_PRODUCT, variables=variables)
+    assert graphql_log_handler.messages == [
+        "saleor.graphql.errors.handled[ERROR].GraphQLError"
+    ]
+    content = get_graphql_content(response, ignore_errors=True)
+    assert len(content["errors"]) == 1
+
+
 def test_fetch_all_products(user_api_client, product):
     response = user_api_client.post_graphql(QUERY_FETCH_ALL_PRODUCTS)
     content = get_graphql_content(response)